Using CSS classes
Added class to the currently active page, called selected.

This class, and any elements undergoing a hover event, will be styled
with a dark green text color.

Also styled the body, it should be white anyway but just to make sure we
don't get overridden. Also set a default text color of a dark gray. More
specific selectors will not be affected.