Text updates
Getting some new fonts we can use from Google Fonts.

These fonts will be utilized in main.css. Right now, we are just setting
the h1 font to use Changa One when available, and sans-serif if we run
into any issues. There is a 15px vertical margin on the h1 element, and
its font size is using the relative until of 1.75 em's.