Styled the gallery
Gallery now looks pretty awesome :)

Made sure that all margins/padding have been removed and got rid of the
bulletpoints.

Also limited each image to a maximum width of 100% of their parent -
they will no longer overflow.

Each list item in the gallery is set to float left. This means that the
elements will float to the left of the parent. All elements also have a
width of 45%, which means we can fit two elements in each row with 10%
spare. This 10% is used by the 2.5% margin, on both sides of each
element (5% for each element) of which 5%*2=10%, so the full 100% of
screen width is used.

Also set a background color for each gallery item as a light gray.
Also set a text color of a darkish gray, which should not be used (It
indicates text that has not been linked).