Improved styling for Contact page
Added an image icon beside each Contact method. The ../ means go up a
level in the File Structure.

Removed bullets, removed any padding, removed any margin and decreased
font size to 0.9em.
Then, where we're more specific, set the padding to 30px on the left
with a 10px margin on the top/bottom.

Also, set display to be block, with a minimum height of 20px and a size
of 20px x 20px. Make sure the background does not repeat.

Added appropriate Google+ logos as well. Styled according to Brand
Guidelines, in accordance so site theme.